new uvs2 releases : uppsrc-2593  tutorial-38  examples-141  reference-113
git-svn-id: svn://ultimatepp.org/upp/trunk@298 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/HeapImp.h
@@ -100,13 +100,13 @@ struct Heap {
 	static Heap  aux;
 	
 #ifdef HEAPDBG
-	static void  DbgFreeFill(void *ptr, int size);
-	static void  DbgFreeCheck(void *ptr, int size);
+	static void  DbgFreeFill(void *ptr, size_t size);
+	static void  DbgFreeCheck(void *ptr, size_t size);
 	static void  DbgFreeFillK(void *ptr, int k);
 	static void *DbgFreeCheckK(void *p, int k);
 #else
-	static void  DbgFreeFill(void *ptr, int size) {}
-	static void  DbgFreeCheck(void *ptr, int size) {}
+	static void  DbgFreeFill(void *ptr, size_t size) {}
+	static void  DbgFreeCheck(void *ptr, size_t size) {}
 	static void  DbgFreeFillK(void *ptr, int k) {}
 	static void *DbgFreeCheckK(void *p, int k) { return p; }
 #endif
